requirements.txt
django~=4.1.3
pyTelegramBotAPI
django-apscheduler
whitenoise
gunicorn
dj-database-url
python-binance
django-simpleui
APScheduler~=3.9.1
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
pytz
django-widget-tweaks
django-auto-prefetch==0.51.0
django-allauth==2.28.1
ccxt~=2.4.5
binance
six~=1.16.0
pathlib~=1.0.1
beautifulsoup4~=4.11.1
html2text~=2020.1.16
pathlib~=1.0.1
fake_useragent
undetected_chromedriver
webdriver_manager
more_itertools
packaging
selenium~=4.6.0
python-dateutil~=2.8.2
asgiref~=3.5.2